perf: add seed parameterization to benchmarks
This is for less noisy benchmarks and more granual
benchmark data.

Discussion:
Before this commit, each iteration inside the benchmark would change the random
seed that controls RCSP instance creation. Keeping in mind that the
latency/duration depends a lot of the instance, then results of
benchmarking two algorithms could depends a lot on how many
benchmark iterations were run. Think of two benchmarks, A and B, benchmarking
algo A and B, respectively, both using the same seeds and thus instances. If
A is runs faster it might given an extra instance to run but if
that instance is difficult (slow to solve) then it will make A look
bad.