perf: fewer label passes
This is for faster run times.

For more than 5 sites over 10 fixed random seeds, benchmarks show around 10%
faster run times.

This is easy to revert so I am not going to document the benchmarking.

The effect of this change is instance dependent and in cases where
new states often dominated then it would lead to more computations.
But that might not matter if memory bound, etc.